+++ package.json
@@ -1,39 +1,18 @@
 {
-  "name": "cali",
-  "version": "0.1.0",
-  "description": "AI agent that runs React Native apps",
-  "type": "module",
-  "main": "./dist/index.js",
-  "bin": {
-    "cali": "./dist/cli.js"
-  },
+  "name": "@cali/root",
+  "private": true,
+  "workspaces": [
+    "packages/*"
+  ],
   "scripts": {
     "prepare": "bun run build:tools && bun run build:cli",
-    "build:tools": "bun build ./src/index.ts --outdir ./dist --target node --packages external",
-    "build:cli": "bun build ./src/cli.ts --outdir ./dist --target node",
+    "build:tools": "bun build ./packages/tools/src/index.ts --outdir ./packages/tools/dist --target node --packages external",
+    "build:cli": "bun build ./packages/cali/src/cli.ts --outdir ./packages/cali/dist --target node",
     "release": "release-it"
   },
-  "dependencies": {
-    "@ai-sdk/openai": "^1.0.2",
-    "@clack/prompts": "^0.8.1",
-    "@react-native-community/cli": "^15.1.2",
-    "@react-native-community/cli-config": "^15.1.2",
-    "@react-native-community/cli-platform-android": "^15.1.2",
-    "@react-native-community/cli-platform-apple": "^15.1.2",
-    "@react-native-community/cli-platform-ios": "^15.1.2",
-    "ai": "^4.0.3",
-    "chalk": "^5.3.0",
-    "dedent": "^1.5.3",
-    "diff": "^7.0.0",
-    "dotenv": "^16.4.5",
-    "gradient-string": "^3.0.0",
-    "zod": "^3.23.8"
-  },
   "devDependencies": {
-    "@react-native-community/cli-types": "^15.1.2",
     "@release-it-plugins/workspaces": "^4.2.0",
     "@release-it/conventional-changelog": "^9.0.3",
-    "@types/diff": "^6.0.0",
     "@typescript-eslint/eslint-plugin": "^7.3.1",
     "@typescript-eslint/parser": "^7.3.1",
     "bun-types": "^1.1.33",
@@ -42,7 +21,6 @@
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-prettier": "^5.0.0",
     "eslint-plugin-simple-import-sort": "^12.0.0",
-    "glob": "^11.0.0",
     "prettier": "^3.2.5",
     "release-it": "^17.10.0",
     "typescript": "^5.1.3",
@@ -51,31 +29,6 @@
   "patchedDependencies": {
     "ai@4.0.3": "patches/ai@4.0.3.patch"
   },
-  "author": "Mike Grabowski <mike@callstack.com>",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/callstackincubator/cali.git"
-  },
-  "contributors": [
-    "Oskar Kwasniewski <oskarkwasniewski@icloud.com> (https://github.com/okwasniewski)"
-  ],
   "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/callstackincubator/cali/issues"
-  },
-  "homepage": "https://github.com/callstackincubator/cali/#readme",
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/",
-    "access": "public"
-  },
-  "keywords": [
-    "react-native",
-    "ai",
-    "agent",
-    "vercel ai",
-    "chatgpt"
-  ],
-  "engines": {
-    "node": ">=22"
-  }
+  "version": "0.1.0"
 }
