+++ packages/mcp-server/README.md
@@ -1,0 +1,6 @@
+# @cali/mcp-server
+
+> [!NOTE]
+> This package is not yet released. It is a work in progress.
+
+Model Context Protocol server that allows you to build and work with React Native apps. Under the hood it uses [@cali/tools](./tools/README.md).
