+++ packages/mcp-server/index.ts
@@ -1,0 +1,1 @@
+console.log('Cooming soon! (tm)')
