+++ packages/mcp-server/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@cali/mcp-server",
+  "private": true,
+  "module": "src/index.ts",
+  "version": "0.1.0"
+}
