--- src/index.ts
@@ -1,3 +1,0 @@
-export * as androidTools from './tools/android'
-export * as iosTools from './tools/apple'
-export * as reactNativeTools from './tools/react-native'
